Force devices into the 'pda' view of TRACS
Instead of overriding every navigation action in the WKWebView
I am replacing the word 'site' in the notification and site
urls to read 'pda'. This means the user can still go to full view
if they desire (for gradebook use), but they will get PDA view
whenever they click a native element and go to the WKWebView.